+++ packages/fetch-router/src/lib/route-handlers.ts
@@ -1,4 +1,4 @@
-import type { Params } from '@remix-run/route-pattern'
+import type { Params, RoutePattern } from '@remix-run/route-pattern'
 
 import type { Middleware } from './middleware.ts'
 import type { RequestHandler } from './request-handler.ts'
@@ -29,25 +29,28 @@ export function isRouteHandlersWithMiddleware<T extends RouteMap>(
   )
 }
 
+type PatternLike = string | RoutePattern | Route
+
 /**
- * Infer the route handler type from a route or string.
+ * Infer the route handler type from a route or route pattern.
  */
 // prettier-ignore
-export type InferRouteHandler<T extends Route | string> =
-  T extends Route<infer M, infer P> ? RouteHandler<M, P> :
-  T extends string ? RouteHandler<'ANY', T> :
+export type InferRouteHandler<Method extends RequestMethod | 'ANY', P extends PatternLike> =
+  P extends string ? RouteHandler<Method, P> :
+  P extends RoutePattern<infer S> ? RouteHandler<Method, S> :
+  P extends Route<infer _, infer S> ? RouteHandler<Method, S> :
   never
 
 /**
  * An individual route handler.
  */
-export type RouteHandler<M extends RequestMethod | 'ANY', T extends string> =
-  | RequestHandlerWithMiddleware<M, T>
-  | RequestHandler<M, Params<T>>
+export type RouteHandler<M extends RequestMethod | 'ANY', P extends string> =
+  | RequestHandlerWithMiddleware<M, P>
+  | RequestHandler<M, Params<P>>
 
-type RequestHandlerWithMiddleware<M extends RequestMethod | 'ANY', T extends string> = {
-  use: Middleware<M, Params<T>>[]
-  handler: RequestHandler<M, Params<T>>
+type RequestHandlerWithMiddleware<M extends RequestMethod | 'ANY', P extends string> = {
+  use: Middleware<M, Params<P>>[]
+  handler: RequestHandler<M, Params<P>>
 }
 
 export function isRequestHandlerWithMiddleware<M extends RequestMethod | 'ANY', T extends string>(