+++ packages/fetch-proxy/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`fetch-proxy`](https://github.com/remix-run/remix/tree/main/packages/fetch-proxy). It follows [semantic versioning](https://semver.org/).
 
+## v0.6.0 (2025-10-22)
+
+- BREAKING CHANGE: Removed CommonJS build. This package is now ESM-only. If you need to use this package in a CommonJS project, you will need to use dynamic `import()`.
+
 ## v0.5.0 (2025-07-24)
 
 - Renamed package from `@mjackson/fetch-proxy` to `@remix-run/fetch-proxy`