+++ packages/form-data-parser/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`form-data-parser`](https://github.com/remix-run/remix/tree/main/packages/form-data-parser). It follows [semantic versioning](https://semver.org/).
 
+## v0.12.0 (2025-10-22)
+
+- BREAKING CHANGE: Removed CommonJS build. This package is now ESM-only. If you need to use this package in a CommonJS project, you will need to use dynamic `import()`.
+
 ## v0.11.0 (2025-10-05)
 
 - Make `options` optional in `parseFormData` signature