+++ packages/headers/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`headers`](https://github.com/remix-run/remix/tree/main/packages/headers). It follows [semantic versioning](https://semver.org/).
 
+## v0.14.0 (2025-10-22)
+
+- BREAKING CHANGE: Removed CommonJS build. This package is now ESM-only. If you need to use this package in a CommonJS project, you will need to use dynamic `import()`.
+
 ## v0.13.0 (2025-10-04)
 
 - Drop support for TypeScript < 5.7