+++ packages/lazy-file/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This is the changelog for [`lazy-file`](https://github.com/remix-run/remix/tree/main/packages/lazy-file). It follows [semantic versioning](https://semver.org/).
 
+## v3.6.0 (2025-10-22)
+
+- BREAKING CHANGE: Removed CommonJS build. This package is now ESM-only. If you need to use this package in a CommonJS project, you will need to use dynamic `import()`.
+
 ## v3.5.0 (2025-07-21)
 
 - Renamed package from `@mjackson/lazy-file` to `@remix-run/lazy-file`